% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_history.R
\name{comparison_history}
\alias{comparison_history}
\title{Compare a single card (turn_1) with the history (working memory), 
else take second card}
\usage{
comparison_history(turn_1, set, set_before_turn_1, history)
}
\description{
Compare a single card (turn_1) with the history (working memory), 
else take second card
}
\author{
Markus Huff
}
