% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_turn_2.R
\name{comparison_turn_2}
\alias{comparison_turn_2}
\title{Compare two cards}
\usage{
comparison_turn_2(turn_1, turn_2, set, set_before_turn_1, history)
}
\description{
Compare two cards
}
\author{
Markus Huff
}

