% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concentration_game.R
\name{concentration_game}
\alias{concentration_game}
\title{Simulation of the concentration game}
\usage{
concentration_game(n = 16, num_simulations = 10, capacity = 4,
  method = "random")
}
\arguments{
\item{n}{Number of card pairs of the game}

\item{num_simulations}{Number of simulations of the game}

\item{capacity}{Working memory capacity}

\item{method}{Characteristic of the working memory ("random","fifo","lifo")}
}
\description{
Simulation of the concentration game
}
\author{
Markus Huff
}
