% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_result.R
\name{print_result}
\alias{print_result}
\title{Print the current result of the simulation}
\usage{
print_result(set, set_before_turn_1)
}
\description{
Print the current result of the simulation
}
\author{
Markus Huff
}
