% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_history.R
\name{reduce_history}
\alias{reduce_history}
\title{Reduce history to working memory capacity}
\usage{
reduce_history(history, capacity, method)
}
\description{
Reduce history to working memory capacity
}
\author{
Markus Huff
}

