% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_missing_rows.R
\name{return_missing_rows}
\alias{return_missing_rows}
\title{Compare two data frames and return those rows that are missing in the second one}
\usage{
return_missing_rows(set_before_turn_1, set)
}
\description{
Compare two data frames and return those rows that are missing in the second one
}
\author{
Markus Huff
}

